refactor(night): extract 4 executors from useActionOrchestrator (C09)
Extract reveal, wolfVote, actionSubmit (magicianFirst +
actionConfirm), and skip intent handlers into dedicated
executor files registered via the executor registry.
useActionOrchestrator shrinks from 785 to 568 lines.

New files:
- revealExecutor.ts (reveal intent)
- wolfVoteExecutor.ts (wolfVote intent)
- actionSubmitExecutor.ts (magicianFirst + actionConfirm)
- skipExecutor.ts (skip intent)

Updated delegationSeatIdentity contract tests to reference
executor files instead of useActionOrchestrator for migrated
intent handlers. Removed unused imports from the orchestrator.
